Change builtins arithmetics types to Double and Long

+++ src/main/java/ru/nsu/fit/amdp/lisp_machine/runtime/expressions/builtins/comparators/LispLess.java
@@ -14,13 +14,13 @@ public class LispLess extends ArithmeticOperation {
         if(args.size() != 2)
             throw new IllegalArgumentException("Incorrect amount of args for less");
 
-        if (args.stream().anyMatch(a -> ((LispObject) a).self() instanceof Float)) {
-            var res = ((Number) ((LispObject) args.get(0)).self()).floatValue()
-                    < ((Number) ((LispObject) args.get(1)).self()).floatValue();
+        if (args.stream().anyMatch(a -> ((LispObject) a).self() instanceof Double)) {
+            var res = ((Number) ((LispObject) args.get(0)).self()).doubleValue()
+                    < ((Number) ((LispObject) args.get(1)).self()).doubleValue();
             return new LispObject(res);
         } else {
-            var res = ((Number) ((LispObject) args.get(0)).self()).intValue()
-                    < ((Number) ((LispObject) args.get(1)).self()).intValue();
+            var res = ((Number) ((LispObject) args.get(0)).self()).longValue()
+                    < ((Number) ((LispObject) args.get(1)).self()).longValue();
             return new LispObject(res);
         }
     }
